// stake-web/src/app/modules/settings/settings.types.ts
export interface PartnerRegistration {
    status?: number;
    _id?: string;
    userId?: string;
    time?: string;
    phone?: string;
    email?: string;
    telegram?: string;
    zalo?: string;
    facebook?: string;
    __v?: number;
    address?: string;
    username?: string;
    name?: string;
}